pyproject.toml
[project]
name = "trader-v1"
version = "0.1.0"
description = "Stock trading prediction system"
authors = [
    {name = "MotiTheWizerd",email = "motielmakayes@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
package-mode = false
dependencies = [
    "rich (>=14.0.0,<15.0.0)",
    "yfinance (>=0.2.65,<0.3.0)",
    "pandas (>=1.5.0,<2.0.0)",
    "numpy (>=1.24.0,<2.0.0)",
    "google-adk (>=1.5.0,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "pandas-market-calendars (>=5.1.1,<6.0.0)",
    "finnhub-python (>=2.4.24,<3.0.0)",
    "typer[all] (>=0.16.0,<0.17.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "fastapi (>=0.116.0,<0.117.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "click (>=8.2.1,<9.0.0)",
    "h11 (>=0.16.0,<0.17.0)",
    "httptools (>=0.6.4,<0.7.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.3,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

